create separate files for dom helpers
