+++ scripts/core/helpers/dom/isElementInViewport.js
@@ -1,0 +1,8 @@
+export function isElementInViewport(element) {
+    var rect = element.getBoundingClientRect();
+
+    return rect.top >= 0
+        && rect.left >= 0
+        && rect.bottom <= (window.innerHeight || document.documentElement.clientHeight)
+        && rect.right <= (window.innerWidth || document.documentElement.clientWidth);
+}
