+++ scripts/core/helpers/dom/querySelectorParent.js
@@ -1,0 +1,9 @@
+export function querySelectorParent(element, selector) {
+    var currentParent = element.parentElement;
+
+    while (currentParent != null && currentParent.matches(selector) === false) {
+        currentParent = currentParent.parentElement;
+    }
+
+    return currentParent;
+}
